# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AccountProfile {
  email: string;
  marketingOptIn: boolean;
}

export interface LessonPlan {
  title: string;
  integratedUnitGoal: string;
  contentObjective: string;
  languageObjective: string;
  languageFunctionObjective: string;
  languageFeatureObjective: string;
  keyVocabulary: string[];
  sentenceFrames: string[];
  warmUp: string;
  mainActivity: string;
  speakingActivity: string;
  exitTicket: string;
  teacherNotes: string;
  scaffoldPlan: string;
  scaffoldFadingPlan: string;
  formativeAssessment: string;
  sourcesUsed: string[];
}

export interface AccountLesson {
  id: string;
  savedAt: string;
  title: string;
  gradeLevel: string;
  languageSupportLevel: string;
  topic: string;
  unitProfile?: string;
  lesson: LessonPlan;
}

export interface AccountLessonsResponse {
  account: AccountProfile;
  lessons: AccountLesson[];
}

export interface SaveAccountLessonBody {
  id: string;
  title: string;
  gradeLevel: string;
  languageSupportLevel: string;
  topic: string;
  unitProfile?: string;
  lesson: LessonPlan;
  marketingOptIn?: boolean;
}

export interface SaveAccountLessonResponse {
  account: AccountProfile;
  lesson: AccountLesson;
}

export type GenerateLessonPlanBodyGradeLevel =
  (typeof GenerateLessonPlanBodyGradeLevel)[keyof typeof GenerateLessonPlanBodyGradeLevel];

export const GenerateLessonPlanBodyGradeLevel = {
  Grade_1: "Grade 1",
  Grade_2: "Grade 2",
  Grade_3: "Grade 3",
  Grade_4: "Grade 4",
  Grade_5: "Grade 5",
} as const;

export type GenerateLessonPlanBodyLanguageSupportLevel =
  (typeof GenerateLessonPlanBodyLanguageSupportLevel)[keyof typeof GenerateLessonPlanBodyLanguageSupportLevel];

export const GenerateLessonPlanBodyLanguageSupportLevel = {
  NUMBER_1: "1",
  NUMBER_2: "2",
  NUMBER_3: "3",
  NUMBER_4: "4",
  NUMBER_5: "5",
  NUMBER_6: "6",
} as const;

/**
 * Optional canonical curriculum unit used to ground the lesson
 */
export type GenerateLessonPlanBodyUnitProfile =
  (typeof GenerateLessonPlanBodyUnitProfile)[keyof typeof GenerateLessonPlanBodyUnitProfile];

export const GenerateLessonPlanBodyUnitProfile = {
  "Grade_4_Discipline-Based_Writing": "Grade 4 Discipline-Based Writing",
} as const;

export interface GenerateLessonPlanBody {
  /** Rough planning notes from the teacher */
  notes: string;
  gradeLevel: GenerateLessonPlanBodyGradeLevel;
  languageSupportLevel: GenerateLessonPlanBodyLanguageSupportLevel;
  /** Topic or subject for the lesson */
  topic: string;
  /** Optional canonical curriculum unit used to ground the lesson */
  unitProfile?: GenerateLessonPlanBodyUnitProfile;
  /** Optional school access code for private-beta authorization */
  accessCode?: string;
}

export type GenerateClassroomSupportBodyGradeLevel =
  (typeof GenerateClassroomSupportBodyGradeLevel)[keyof typeof GenerateClassroomSupportBodyGradeLevel];

export const GenerateClassroomSupportBodyGradeLevel = {
  Grade_1: "Grade 1",
  Grade_2: "Grade 2",
  Grade_3: "Grade 3",
  Grade_4: "Grade 4",
  Grade_5: "Grade 5",
} as const;

export type GenerateClassroomSupportBodyLanguageSupportLevel =
  (typeof GenerateClassroomSupportBodyLanguageSupportLevel)[keyof typeof GenerateClassroomSupportBodyLanguageSupportLevel];

export const GenerateClassroomSupportBodyLanguageSupportLevel = {
  NUMBER_1: "1",
  NUMBER_2: "2",
  NUMBER_3: "3",
  NUMBER_4: "4",
  NUMBER_5: "5",
  NUMBER_6: "6",
} as const;

export interface GenerateClassroomSupportBody {
  /** What the students need help with right now */
  need: string;
  gradeLevel: GenerateClassroomSupportBodyGradeLevel;
  languageSupportLevel: GenerateClassroomSupportBodyLanguageSupportLevel;
  /** Optional school access code for private-beta authorization */
  accessCode?: string;
}

export interface ClassroomSupport {
  simpleExplanation: string;
  keyVocabulary: string[];
  sentenceFrames: string[];
  quickActivity: string;
  extensionQuestion: string;
  teacherMove: string;
}

export interface ErrorResponse {
  error: string;
}

export type ValidateAccessCodeBody = {
  /** Optional school access code when private-beta access is enabled */
  accessCode?: string;
};

export type ValidateAccessCode200 = {
  valid: boolean;
};