# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: lessonPlan
    description: Lesson plan generation operations
  - name: classroomCopilot
    description: Classroom Copilot quick EAL support operations
  - name: account
    description: Verified account and saved lesson operations
paths:
  /access/validate:
    post:
      operationId: validateAccessCode
      summary: Validate a verified account and, when enabled, a private-beta access code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                accessCode:
                  type: string
                  description: Optional school access code when private-beta access is enabled
      responses:
        "200":
          description: Valid code; no generation quota consumed
          content:
            application/json:
              schema:
                type: object
                required: [valid]
                properties:
                  valid:
                    type: boolean
        "401":
          description: Invalid or missing access code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /account/lessons:
    get:
      operationId: listAccountLessons
      tags: [account]
      summary: List saved lessons for the verified account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account profile and saved lessons
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountLessonsResponse"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: saveAccountLesson
      tags: [account]
      summary: Save or update a lesson for the verified account
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveAccountLessonBody"
      responses:
        "200":
          description: Saved lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SaveAccountLessonResponse"
        "400":
          description: Invalid lesson
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Sign-in required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /lesson-plan/generate:
    post:
      operationId: generateLessonPlan
      tags: [lessonPlan]
      summary: Generate a lesson plan from teacher notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateLessonPlanBody"
      responses:
        "200":
          description: Generated lesson plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LessonPlan"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /classroom-copilot/generate:
    post:
      operationId: generateClassroomSupport
      tags: [classroomCopilot]
      summary: Generate quick EAL classroom support
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateClassroomSupportBody"
      responses:
        "200":
          description: Generated classroom support
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassroomSupport"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    AccountProfile:
      type: object
      properties:
        email:
          type: string
          format: email
        marketingOptIn:
          type: boolean
      required:
        - email
        - marketingOptIn
    AccountLesson:
      type: object
      properties:
        id:
          type: string
          format: uuid
        savedAt:
          type: string
          format: date-time
        title:
          type: string
        gradeLevel:
          type: string
        languageSupportLevel:
          type: string
        topic:
          type: string
        unitProfile:
          type: string
        lesson:
          $ref: "#/components/schemas/LessonPlan"
      required:
        - id
        - savedAt
        - title
        - gradeLevel
        - languageSupportLevel
        - topic
        - lesson
    AccountLessonsResponse:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountProfile"
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/AccountLesson"
      required:
        - account
        - lessons
    SaveAccountLessonBody:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        gradeLevel:
          type: string
        languageSupportLevel:
          type: string
        topic:
          type: string
        unitProfile:
          type: string
        lesson:
          $ref: "#/components/schemas/LessonPlan"
        marketingOptIn:
          type: boolean
      required:
        - id
        - title
        - gradeLevel
        - languageSupportLevel
        - topic
        - lesson
    SaveAccountLessonResponse:
      type: object
      properties:
        account:
          $ref: "#/components/schemas/AccountProfile"
        lesson:
          $ref: "#/components/schemas/AccountLesson"
      required:
        - account
        - lesson
    GenerateLessonPlanBody:
      type: object
      properties:
        notes:
          type: string
          description: Rough planning notes from the teacher
        gradeLevel:
          type: string
          enum: ["Grade 1", "Grade 2", "Grade 3", "Grade 4", "Grade 5"]
        languageSupportLevel:
          type: string
          enum: ["1", "2", "3", "4", "5", "6"]
        topic:
          type: string
          description: Topic or subject for the lesson
        unitProfile:
          type: string
          enum: ["Grade 4 Discipline-Based Writing"]
          description: Optional canonical curriculum unit used to ground the lesson
        accessCode:
          type: string
          description: Optional school access code for private-beta authorization
      required:
        - notes
        - gradeLevel
        - languageSupportLevel
        - topic
    LessonPlan:
      type: object
      properties:
        title:
          type: string
        integratedUnitGoal:
          type: string
        contentObjective:
          type: string
        languageObjective:
          type: string
        languageFunctionObjective:
          type: string
        languageFeatureObjective:
          type: string
        keyVocabulary:
          type: array
          items:
            type: string
        sentenceFrames:
          type: array
          items:
            type: string
        warmUp:
          type: string
        mainActivity:
          type: string
        speakingActivity:
          type: string
        exitTicket:
          type: string
        teacherNotes:
          type: string
        scaffoldPlan:
          type: string
        scaffoldFadingPlan:
          type: string
        formativeAssessment:
          type: string
        sourcesUsed:
          type: array
          items:
            type: string
      required:
        - title
        - integratedUnitGoal
        - contentObjective
        - languageObjective
        - languageFunctionObjective
        - languageFeatureObjective
        - keyVocabulary
        - sentenceFrames
        - warmUp
        - mainActivity
        - speakingActivity
        - exitTicket
        - teacherNotes
        - scaffoldPlan
        - scaffoldFadingPlan
        - formativeAssessment
        - sourcesUsed
    GenerateClassroomSupportBody:
      type: object
      properties:
        need:
          type: string
          description: What the students need help with right now
        gradeLevel:
          type: string
          enum: ["Grade 1", "Grade 2", "Grade 3", "Grade 4", "Grade 5"]
        languageSupportLevel:
          type: string
          enum: ["1", "2", "3", "4", "5", "6"]
        accessCode:
          type: string
          description: Optional school access code for private-beta authorization
      required:
        - need
        - gradeLevel
        - languageSupportLevel
    ClassroomSupport:
      type: object
      properties:
        simpleExplanation:
          type: string
        keyVocabulary:
          type: array
          items:
            type: string
        sentenceFrames:
          type: array
          items:
            type: string
        quickActivity:
          type: string
        extensionQuestion:
          type: string
        teacherMove:
          type: string
      required:
        - simpleExplanation
        - keyVocabulary
        - sentenceFrames
        - quickActivity
        - extensionQuestion
        - teacherMove
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error