# lib/api-zod/src/generated/types/accountLesson.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonPlan } from "./lessonPlan";

export interface AccountLesson {
  id: string;
  savedAt: Date;
  title: string;
  gradeLevel: string;
  languageSupportLevel: string;
  topic: string;
  unitProfile?: string;
  lesson: LessonPlan;
}