# lib/api-zod/src/generated/types/accountLessonsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccountLesson } from "./accountLesson";
import type { AccountProfile } from "./accountProfile";

export interface AccountLessonsResponse {
  account: AccountProfile;
  lessons: AccountLesson[];
}